requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.8
aphantasia @ git+https://github.com/pixray/aphantasia@7e6b3bb4519e498f97670c957ba55588f7aa95e7
async-timeout==4.0.2
attrs==21.4.0
Authlib==1.0.1
braceexpand==0.1.7
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
clip @ git+https://github.com/openai/CLIP.git@b4ae44927b78d0093b556e3ce43cbdcff422017a
cryptography==36.0.2
cycler==0.11.0
easydict==1.9
ecdsa==0.17.0
einops==0.4.1
et-xmlfile==1.1.0
Flask==2.0.2
Flask-Cors==3.0.10
frozenlist==1.3.0
fsspec==2022.3.0
ftfy==6.1.1
google-auth==2.6.5
google-auth-oauthlib==0.4.6
grpcio==1.44.0
idna==3.3
imageio==2.17.0
imageio-ffmpeg==0.4.7
importlib-metadata==4.11.3
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.4.2
kornia==0.6.1
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.3.4
multidict==6.0.2
numpy==1.21.4
oauthlib==3.2.0
omegaconf==2.1.2
openpyxl==3.0.9
packaging==21.3
pandas==1.3.4
pandas-stubs==1.2.0.39
perlin-numpy @ git+https://github.com/pvigier/perlin-numpy@6f077f811f5708e504732c26dee8f2015b95da0c
Pillow==9.1.0
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.2
PyJWT==2.3.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytorch-lightning==1.6.1
pytorch-ranger==0.1.1
pytz==2021.3
PyYAML==6.0
regex==2022.3.15
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.0
six==1.16.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch==1.9.0+cu111
torch-optimizer==0.3.0
torchmetrics==0.8.0
torchtext==0.10.0
torchvision==0.10.0+cu111
tqdm==4.62.3
typing_extensions==4.2.0
urllib3==1.26.7
waitress==2.0.0
wcwidth==0.2.5
Werkzeug==2.0.2
yarl==1.7.2
zipp==3.8.0
